test(xsnap): factor options(io) out of xsnap tests
options() was copied between test scripts because importing from a
test script runs its tests and because options() closed over ambient
authority. A new message-tool.js module where io access is passed
explicitly to optins() lets us share, as well as:

  - reduce scope of test-xsnap.js to just the xsnap API
    - move JS correctness tests to their own file, test-xs-js.js
    - move resource exhaustion tests to test-xs-limits.js
  - add @ts-check to facilitate refactoring
    - x.map(JSON.parse) -> x.map(s => JSON.parse(s))
  - group imports: ava, platform, code under test, test modules
  - share just encode / decode methods (encoder / decoder have mutable state)
  - importModuleUrl -> importMetaUrl (again)
  - fix gc test to call gc() rather than print() in the start
    compartment
  - when testing print(), let the output be more clear about where
    it's coming from
  - use t.log, not console.log, for property exhaustion testing
  - factor out asset loader
